refactor(editor): TipTapEditor sous 300 lignes — éclatement complet (H13)
859 -> 186 l. Script -> useTipTapEditor.js ; bubble menu -> EditorBubbleMenu.vue ; modale plein ecran -> TipTapFullscreenModal.vue ; typographie ProseMirror partagee -> partial _tiptap-content.scss ; chrome modale -> partial _tiptap-fullscreen-modal.scss. API (modelValue/placeholder, update:modelValue) inchangee. CSS verbatim, dark mode preserve. Tests parite verts (stubs inchanges) + nouveaux. Dette preexistante conservee : <bubble-menu> non importe/enregistre (rendu inerte) — documentee.
