+++ src/assets/styles/_tiptap-content.scss
@@ -1,0 +1,223 @@
+// Typographie ProseMirror partagée de l'éditeur riche (#G1 ≤300, décompo TipTapEditor).
+// Extraite VERBATIM du <style scoped> de TipTapEditor.vue (ancien bloc « TYPOGRAPHY
+// (Partagé) », lignes ~498-710) : ces règles `:deep(.ProseMirror …)` s'appliquent
+// à l'identique au contenu de l'éditeur en mode NORMAL et au contenu de la modale
+// plein écran. Mutualisées en partial puis `@use`-ées (scoped) dans TipTapEditor.vue
+// ET TipTapFullscreenModal.vue, chacun les scopant à son propre sous-arbre.
+
+/* ===========================
+   TYPOGRAPHY (Partagé)
+   =========================== */
+:deep(.ProseMirror h1) {
+  font-size: 2.25rem;
+  font-weight: 700;
+  margin-top: 2rem;
+  margin-bottom: 1rem;
+  color: var(--text-primary);
+  line-height: 1.2;
+}
+
+:deep(.ProseMirror h2) {
+  font-size: 1.75rem;
+  font-weight: 600;
+  margin-top: 1.75rem;
+  margin-bottom: 0.875rem;
+  color: var(--text-primary);
+  line-height: 1.3;
+}
+
+:deep(.ProseMirror h3) {
+  font-size: 1.375rem;
+  font-weight: 600;
+  margin-top: 1.5rem;
+  margin-bottom: 0.75rem;
+  color: var(--text-primary);
+  line-height: 1.4;
+}
+
+:deep(.ProseMirror p) {
+  margin-top: 0.75rem;
+  margin-bottom: 0.75rem;
+}
+
+/* Lists */
+:deep(.ProseMirror ul),
+:deep(.ProseMirror ol) {
+  padding-left: 2rem;
+  margin-top: 0.75rem;
+  margin-bottom: 0.75rem;
+}
+
+:deep(.ProseMirror li) {
+  margin-top: 0.375rem;
+  margin-bottom: 0.375rem;
+}
+
+/* Task list */
+:deep(.ProseMirror ul[data-type="taskList"]) {
+  list-style: none;
+  padding-left: 0;
+}
+
+:deep(.ProseMirror ul[data-type="taskList"] li) {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.75rem;
+}
+
+:deep(.ProseMirror ul[data-type="taskList"] li input[type="checkbox"]) {
+  margin-top: 0.5rem;
+  cursor: pointer;
+  width: 18px;
+  height: 18px;
+}
+
+/* Code */
+:deep(.ProseMirror code) {
+  background: rgba(0, 0, 0, 0.1);
+  padding: 0.25em 0.5em;
+  border-radius: 4px;
+  font-size: 0.9em;
+  font-family: 'Courier New', monospace;
+  color: var(--text-primary);
+}
+
+[data-theme="dark"] :deep(.ProseMirror code) {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+@media (prefers-color-scheme: dark) {
+  :deep(.ProseMirror code) {
+    background: rgba(255, 255, 255, 0.1);
+  }
+}
+
+:deep(.ProseMirror pre) {
+  background: rgba(0, 0, 0, 0.05);
+  border-radius: 8px;
+  padding: 1.25rem;
+  margin: 1.25rem 0;
+  overflow-x: auto;
+  border: 2px solid var(--border-color);
+}
+
+[data-theme="dark"] :deep(.ProseMirror pre) {
+  background: rgba(255, 255, 255, 0.05);
+}
+
+@media (prefers-color-scheme: dark) {
+  :deep(.ProseMirror pre) {
+    background: rgba(255, 255, 255, 0.05);
+  }
+}
+
+:deep(.ProseMirror pre code) {
+  background: none;
+  padding: 0;
+  font-size: 0.9rem;
+}
+
+/* Blockquote */
+:deep(.ProseMirror blockquote) {
+  border-left: 5px solid var(--color-primary, #10b981);
+  padding-left: 1.5rem;
+  margin: 1.25rem 0;
+  color: var(--text-secondary);
+  font-style: italic;
+  background: rgba(16, 185, 129, 0.05);
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  border-radius: 0 8px 8px 0;
+}
+
+/* Horizontal Rule */
+:deep(.ProseMirror hr) {
+  border: none;
+  border-top: 3px solid var(--border-color);
+  margin: 2rem 0;
+}
+
+/* Links */
+:deep(.ProseMirror a) {
+  color: var(--color-primary, #10b981);
+  text-decoration: underline;
+  cursor: pointer;
+  transition: opacity 0.2s ease;
+}
+
+:deep(.ProseMirror a:hover) {
+  opacity: 0.7;
+}
+
+/* Images */
+:deep(.ProseMirror img) {
+  max-width: 100%;
+  height: auto;
+  border-radius: 10px;
+  margin: 1.5rem 0;
+  display: block;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+/* YouTube video */
+:deep(.ProseMirror iframe) {
+  max-width: 100%;
+  border-radius: 10px;
+  margin: 1.5rem 0;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+/* Tables */
+:deep(.ProseMirror table) {
+  border-collapse: collapse;
+  margin: 1.5rem 0;
+  width: 100%;
+  overflow: hidden;
+  border: 2px solid var(--border-color);
+  border-radius: 10px;
+}
+
+:deep(.ProseMirror table td),
+:deep(.ProseMirror table th) {
+  border: 1px solid var(--border-color);
+  padding: 12px 16px;
+  vertical-align: top;
+  position: relative;
+}
+
+:deep(.ProseMirror table th) {
+  background: var(--bg-secondary);
+  font-weight: 600;
+  text-align: left;
+}
+
+:deep(.ProseMirror table .selectedCell) {
+  background: rgba(16, 185, 129, 0.15);
+}
+
+/* Placeholder */
+:deep(.ProseMirror p.is-editor-empty:first-child::before) {
+  content: attr(data-placeholder);
+  float: left;
+  color: var(--text-muted, #9ca3af);
+  pointer-events: none;
+  height: 0;
+  font-style: italic;
+}
+
+/* Text alignment */
+:deep(.ProseMirror [style*="text-align: left"]) {
+  text-align: left;
+}
+
+:deep(.ProseMirror [style*="text-align: center"]) {
+  text-align: center;
+}
+
+:deep(.ProseMirror [style*="text-align: right"]) {
+  text-align: right;
+}
+
+:deep(.ProseMirror [style*="text-align: justify"]) {
+  text-align: justify;
+}
