+++ src/assets/styles/_tiptap-fullscreen-modal.scss
@@ -1,0 +1,266 @@
+// Chrome de la modale plein écran de l'éditeur riche (#G1 ≤300, décompo TipTapEditor).
+// Extrait VERBATIM du <style scoped> de TipTapEditor.vue (overlay, container +
+// @keyframes modalSlideIn, header, btn-close-modal, modal-toolbar, modal-content +
+// :deep(.ProseMirror) modale, footer, fullscreen-hint, transition modal-fade,
+// scrollbar .modal-content-wrapper, @media responsive). Mutualisé en partial pour
+// garder le SFC TipTapFullscreenModal.vue ≤300 lignes ; `@use`-é (scoped) dedans.
+
+/* ===========================
+   MODAL OVERLAY (90%)
+   =========================== */
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.8);
+  z-index: 99999;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 20px;
+  backdrop-filter: blur(4px);
+}
+
+/* ===========================
+   MODAL CONTAINER (La carte popup)
+   =========================== */
+.modal-container {
+  width: 90vw;
+  height: 90vh;
+  max-width: 1600px;
+  background: var(--card-bg);
+  border-radius: 16px;
+  box-shadow: 0 25px 50px -12px rgba(0, 0, 0, 0.5);
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  animation: modalSlideIn 0.3s ease-out;
+}
+
+@keyframes modalSlideIn {
+  from {
+    opacity: 0;
+    transform: scale(0.95) translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: scale(1) translateY(0);
+  }
+}
+
+/* ===========================
+   MODAL HEADER
+   =========================== */
+.modal-header {
+  background: var(--bg-secondary);
+  border-bottom: 2px solid var(--border-color);
+  padding: 20px 30px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-shrink: 0;
+}
+
+.modal-title {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: var(--text-primary);
+  margin: 0;
+}
+
+.btn-close-modal {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 10px 20px;
+  background: var(--color-primary, #10b981);
+  color: white;
+  border: none;
+  border-radius: 8px;
+  font-size: 0.95rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.btn-close-modal:hover {
+  background: #059669;
+  transform: translateY(-2px);
+  box-shadow: 0 8px 16px rgba(16, 185, 129, 0.3);
+}
+
+.close-icon {
+  font-size: 1.25rem;
+  font-weight: bold;
+}
+
+/* ===========================
+   MODAL TOOLBAR
+   =========================== */
+.modal-toolbar {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 6px;
+  padding: 16px 30px;
+  background: var(--bg-secondary);
+  border-bottom: 2px solid var(--border-color);
+  overflow-x: auto;
+  overflow-y: hidden;
+  flex-shrink: 0;
+}
+
+/* ===========================
+   MODAL CONTENT
+   =========================== */
+.modal-content-wrapper {
+  flex: 1;
+  overflow-y: auto;
+  overflow-x: hidden;
+  background: var(--bg-primary);
+}
+
+.modal-editor-content {
+  height: 100%;
+}
+
+.modal-editor-content :deep(.ProseMirror) {
+  padding: 40px 60px;
+  min-height: 100%;
+  max-width: 1200px;
+  margin: 0 auto;
+  background-color: var(--card-bg);
+  color: var(--text-primary);
+  outline: none;
+  font-size: 1.05rem;
+  line-height: 1.75;
+}
+
+/* ===========================
+   MODAL FOOTER
+   =========================== */
+.modal-footer {
+  padding: 16px 30px;
+  background: var(--bg-secondary);
+  border-top: 2px solid var(--border-color);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 0.9rem;
+  color: var(--text-secondary);
+  flex-shrink: 0;
+}
+
+.footer-left {
+  display: flex;
+  gap: 20px;
+}
+
+.word-count,
+.char-count {
+  color: var(--text-secondary);
+  font-weight: 500;
+}
+
+.footer-right {
+  display: flex;
+  align-items: center;
+}
+
+.fullscreen-hint {
+  font-size: 0.85rem;
+  color: var(--text-muted);
+}
+
+.fullscreen-hint kbd {
+  background: var(--card-bg);
+  border: 1px solid var(--border-color);
+  border-radius: 4px;
+  padding: 3px 8px;
+  font-family: monospace;
+  font-size: 0.85rem;
+  margin: 0 4px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+/* ===========================
+   ANIMATIONS
+   =========================== */
+.modal-fade-enter-active,
+.modal-fade-leave-active {
+  transition: opacity 0.3s ease;
+}
+
+.modal-fade-enter-from,
+.modal-fade-leave-to {
+  opacity: 0;
+}
+
+/* Scrollbar (part modale du sélecteur combiné d'origine, dédoublée verbatim) */
+.modal-content-wrapper::-webkit-scrollbar {
+  width: 12px;
+}
+
+.modal-content-wrapper::-webkit-scrollbar-track {
+  background: var(--bg-secondary);
+}
+
+.modal-content-wrapper::-webkit-scrollbar-thumb {
+  background: var(--border-color);
+  border-radius: 6px;
+  border: 2px solid var(--bg-secondary);
+}
+
+.modal-content-wrapper::-webkit-scrollbar-thumb:hover {
+  background: var(--text-muted);
+}
+
+/* ===========================
+   RESPONSIVE
+   =========================== */
+@media (max-width: 1024px) {
+  .modal-container {
+    width: 95vw;
+    height: 95vh;
+  }
+
+  .modal-editor-content :deep(.ProseMirror) {
+    padding: 30px 40px;
+  }
+}
+
+@media (max-width: 768px) {
+  .modal-header {
+    padding: 16px 20px;
+  }
+
+  .modal-title {
+    font-size: 1.25rem;
+  }
+
+  .btn-close-modal .close-text {
+    display: none;
+  }
+
+  .modal-toolbar {
+    padding: 12px 20px;
+  }
+
+  .modal-editor-content :deep(.ProseMirror) {
+    padding: 20px 24px;
+  }
+
+  .modal-footer {
+    padding: 12px 20px;
+    flex-direction: column;
+    gap: 12px;
+    align-items: flex-start;
+  }
+
+  .toolbar-btn-fullscreen .btn-label {
+    display: none;
+  }
+}
